Add relative path handling to source contentPath (#1)
* Update gitignore for VScode files

* Update to allow relative content path

Update the argument processing to add a new step for 'processSource' which
will look for a relative content path argument and replace it with a
path.resolve fully qualified version to get around the msdeploy limitation
of always using absolute paths. This allows your npm script to use a
relative path for the contentPath.